+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Master
+
+- [Issue #82](https://github.com/netzpirat/guard-jasmine/issues/82): Enable the `:auto` server to detection unicorn, thin and mongrel.
+
 ## 1.8.3 - September 19, 2012
 
 - [Issue #81](https://github.com/netzpirat/guard-jasmine/issues/81): Make the Guardfile template catch `.js.coffee` again.