+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Master
 
+- [Pull #83](https://github.com/netzpirat/guard-jasmine/pull/83): Make it possible to run within mountable engine within a dummy app (using jasminerice). ([@pschyska][])
 - [Issue #82](https://github.com/netzpirat/guard-jasmine/issues/82): Enable the `:auto` server to detection unicorn, thin and mongrel.
 
 ## 1.8.3 - September 19, 2012
@@ -304,6 +305,7 @@
 [@mutru]: https:///github.com/mutru
 [@obrie]: https://github.com/obrie
 [@pushbang]: https://github.com/pushbang
+[@pschyska]: https://github.com/pschyska
 [@richo]: https://github.com/richo
 [@robotarmy]: https://github.com/robotarmy
 [@rymai]: https://github.com/rymai